# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TourFlow AI
  description: 'TourFlow AI API contract — generated from the FastAPI implementation
    (source of truth: backend/app). Regenerate: python -c "from app.main import app;
    ..."'
  version: 1.0.0
paths:
...
  /api/v1/trips:
    post:
      tags:
      - trips
      summary: Create Trip
      operationId: create_trip_api_v1_trips_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - trips
      summary: List Trips
      operationId: list_trips_api_v1_trips_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/TripStatus'
          - type: 'null'
          title: Status
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TripOut'
                title: Response List Trips Api V1 Trips Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/trips/{trip_id}:
    get:
      tags:
      - trips
      summary: Get Trip
      operationId: get_trip_api_v1_trips__trip_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - trips
      summary: Update Trip
      operationId: update_trip_api_v1_trips__trip_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - trips
      summary: Delete Trip
      operationId: delete_trip_api_v1_trips__trip_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/trips/{trip_id}/preferences:
    put:
      tags:
      - trips
      summary: Upsert Preferences
      operationId: upsert_preferences_api_v1_trips__trip_id__preferences_put
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripPreferenceIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripPreferenceOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - trips
      summary: Get Preferences
      operationId: get_preferences_api_v1_trips__trip_id__preferences_get
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/TripPreferenceOut'
                - type: 'null'
                title: Response Get Preferences Api V1 Trips  Trip Id  Preferences
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/trips/{trip_id}/recommendations:
    get:
      tags:
      - ai
      summary: Recommendations
      operationId: recommendations_api_v1_trips__trip_id__recommendations_get
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecommendationOut'
                title: Response Recommendations Api V1 Trips  Trip Id  Recommendations
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/trips/{trip_id}/itinerary-suggestion:
    get:
      tags:
      - ai
      summary: Itinerary Suggestion
      description: AI day-by-day skeleton. Caller must validate before persisting
        as items.
      operationId: itinerary_suggestion_api_v1_trips__trip_id__itinerary_suggestion_get
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/trips/{trip_id}/ask:
    post:
      tags:
      - ai
      summary: Ask Question
      operationId: ask_question_api_v1_trips__trip_id__ask_post
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripQuestionIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RecommendationOut:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        category:
          type: string
          title: Category
          default: ''
        estimated_cost:
          anyOf:
          - type: number
          - type: 'null'
          title: Estimated Cost
        score:
          type: number
          title: Score
          default: 0.0
      type: object
      required:
      - title
      - description
      title: RecommendationOut
...
    TripCreate:
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 2
          title: Title
        destination:
          type: string
          maxLength: 200
          minLength: 2
          title: Destination
        start_date:
          type: string
          format: date
          title: Start Date
        end_date:
          type: string
          format: date
          title: End Date
        budget:
          anyOf:
          - type: number
            minimum: 0.0
          - type: 'null'
          title: Budget
        currency:
          type: string
          maxLength: 3
          minLength: 3
          title: Currency
          default: INR
        travel_style:
          anyOf:
          - type: string
            maxLength: 60
          - type: 'null'
          title: Travel Style
      type: object
      required:
      - title
      - destination
      - start_date
      - end_date
      title: TripCreate
    TripOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        traveler_id:
          type: string
          format: uuid
          title: Traveler Id
        operator_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Operator Id
        coordinator_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Coordinator Id
        title:
          type: string
          title: Title
        destination:
          type: string
          title: Destination
        start_date:
          type: string
          format: date
          title: Start Date
        end_date:
          type: string
          format: date
          title: End Date
        duration_days:
          type: integer
          title: Duration Days
        budget:
          anyOf:
          - type: number
          - type: 'null'
          title: Budget
        currency:
          type: string
          title: Currency
        status:
          $ref: '#/components/schemas/TripStatus'
        travel_style:
          anyOf:
          - type: string
          - type: 'null'
          title: Travel Style
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - traveler_id
      - title
      - destination
      - start_date
      - end_date
      - duration_days
      - currency
      - status
      - created_at
      - updated_at
      title: TripOut
    TripPreferenceIn:
      properties:
        budget:
          anyOf:
          - type: number
            minimum: 0.0
          - type: 'null'
          title: Budget
        accommodation_preference:
          anyOf:
          - type: string
            maxLength: 60
          - type: 'null'
          title: Accommodation Preference
        transportation_preference:
          anyOf:
          - type: string
            maxLength: 60
          - type: 'null'
          title: Transportation Preference
        interests:
          items:
            type: string
          type: array
          title: Interests
        activity_preferences:
          items:
            type: string
          type: array
          title: Activity Preferences
        pace:
          anyOf:
          - type: string
            maxLength: 30
          - type: 'null'
          title: Pace
        travel_style:
          anyOf:
          - type: string
            maxLength: 60
          - type: 'null'
          title: Travel Style
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      title: TripPreferenceIn
    TripPreferenceOut:
      properties:
        budget:
          anyOf:
          - type: number
            minimum: 0.0
          - type: 'null'
          title: Budget
        accommodation_preference:
          anyOf:
          - type: string
            maxLength: 60
          - type: 'null'
          title: Accommodation Preference
        transportation_preference:
          anyOf:
          - type: string
            maxLength: 60
          - type: 'null'
          title: Transportation Preference
        interests:
          items:
            type: string
          type: array
          title: Interests
        activity_preferences:
          items:
            type: string
          type: array
          title: Activity Preferences
        pace:
          anyOf:
          - type: string
            maxLength: 30
          - type: 'null'
          title: Pace
        travel_style:
          anyOf:
          - type: string
            maxLength: 60
          - type: 'null'
          title: Travel Style
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        id:
          type: string
          format: uuid
          title: Id
        trip_id:
          type: string
          format: uuid
          title: Trip Id
      type: object
      required:
      - id
      - trip_id
      title: TripPreferenceOut
    TripQuestionIn:
      properties:
        question:
          type: string
          maxLength: 1000
          minLength: 3
          title: Question
      type: object
      required:
      - question
      title: TripQuestionIn
    TripStatus:
      type: string
      enum:
      - DRAFT
      - PLANNING
      - READY
      - BOOKED
      - IN_PROGRESS
      - DISRUPTED
      - COMPLETED
      - CANCELLED
      title: TripStatus
    TripUpdate:
      properties:
        title:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 2
          - type: 'null'
          title: Title
        destination:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 2
          - type: 'null'
          title: Destination
        start_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Start Date
        end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End Date
        budget:
          anyOf:
          - type: number
            minimum: 0.0
          - type: 'null'
          title: Budget
        travel_style:
          anyOf:
          - type: string
            maxLength: 60
          - type: 'null'
          title: Travel Style
        status:
          anyOf:
          - $ref: '#/components/schemas/TripStatus'
          - type: 'null'
        coordinator_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Coordinator Id
      type: object
      title: TripUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer